requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
Django==5.0.1
djangorestframework==3.15.2
gunicorn==21.2.0
psycopg2==2.9.9
XlsxWriter==3.2.0